backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
openai==1.51.0
python-dotenv==1.0.1
python-multipart==0.0.9
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
pgvector==0.3.4
pydantic==2.9.2
pypdf==5.0.1
python-docx==1.1.2
